LayoutTests/editing/inserting/insert-line-break-whitespace-pre-line-expected.txt

| <span>
|   contenteditable=""
|   style="display: block; white-space: pre-line;"
|   "\n"
|   "<#selection-caret>\n"
